lc0988: Smallest Strign Starting from Leaf
