+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/depth_first_search/tree/summary.txt
@@ -31,6 +31,7 @@ References:
 2   0951    https://leetcode.com/problems/flip-equivalent-binary-trees/
 2   0958    https://leetcode.com/problems/check-completeness-of-a-binary-tree/
 2   0971    https://leetcode.com/problems/flip-binary-tree-to-match-preorder-traversal/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   0998    https://leetcode.com/problems/maximum-binary-tree-ii/
 2   1022    https://leetcode.com/problems/sum-of-root-to-leaf-binary-numbers/
 2   1026    https://leetcode.com/problems/maximum-difference-between-node-and-ancestor/
