+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/depth_first_search/tree/traversal/traversal_postorder.txt
@@ -13,6 +13,7 @@ References:
 2   0590    https://leetcode.com/problems/n-ary-tree-postorder-traversal/
 2   0669    https://leetcode.com/problems/trim-a-binary-search-tree/
 2   0814    https://leetcode.com/problems/binary-tree-pruning/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   1026    https://leetcode.com/problems/maximum-difference-between-node-and-ancestor/
 2   1302    https://leetcode.com/problems/deepest-leaves-sum/
