+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/depth_first_search/tree/traversal/traversal_preorder.txt
@@ -10,6 +10,7 @@ References:
 
 2   0589    https://leetcode.com/problems/n-ary-tree-preorder-traversal/
 2   0637    https://leetcode.com/problems/average-of-levels-in-binary-tree/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   1022    https://leetcode.com/problems/sum-of-root-to-leaf-binary-numbers/
 2   1026    https://leetcode.com/problems/maximum-difference-between-node-and-ancestor/
 2   1302    https://leetcode.com/problems/deepest-leaves-sum/
