+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/string/search/summary.txt
@@ -8,6 +8,7 @@ References:
 
 2   0392    https://leetcode.com/problems/is-subsequence/
 2   0433    https://leetcode.com/problems/minimum-genetic-mutation/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   1047    https://leetcode.com/problems/remove-all-adjacent-duplicates-in-string/
 
 3   0247    https://leetcode.com/problems/strobogrammatic-number-ii/
