+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/string/summary.txt
@@ -32,6 +32,7 @@ References:
 2   0859    https://leetcode.com/problems/buddy-strings/
 2   0953    https://leetcode.com/problems/verifying-an-alien-dictionary/
 2   0984    https://leetcode.com/problems/string-without-aaa-or-bbb/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   1047    https://leetcode.com/problems/remove-all-adjacent-duplicates-in-string/
 2   1071    https://leetcode.com/problems/greatest-common-divisor-of-strings/
 2   1578    https://leetcode.com/problems/minimum-deletion-cost-to-avoid-repeating-letters/
