+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/tree/by_data_structure/leaf.txt
@@ -10,6 +10,7 @@ References:
 
 2   0366    https://leetcode.com/problems/find-leaves-of-binary-tree/
 2   0404    https://leetcode.com/problems/sum-of-left-leaves/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   1022    https://leetcode.com/problems/sum-of-root-to-leaf-binary-numbers/
 2   1302    https://leetcode.com/problems/deepest-leaves-sum/
