+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/tree/by_data_structure/path.txt
@@ -14,6 +14,7 @@ References:
 ### Path Sum
 
 2   0112    https://leetcode.com/problems/path-sum/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 
 3   0064    https://leetcode.com/problems/minimum-path-sum/
 3   0113    https://leetcode.com/problems/path-sum-ii/
