+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/tree/traversal/summary.txt
@@ -22,6 +22,7 @@ References:
 2   0701    https://leetcode.com/problems/insert-into-a-binary-search-tree/
 2   0814    https://leetcode.com/problems/binary-tree-pruning/
 2   0958    https://leetcode.com/problems/check-completeness-of-a-binary-tree/
+2   0988    https://leetcode.com/problems/smallest-string-starting-from-leaf/
 2   0998    https://leetcode.com/problems/maximum-binary-tree-ii/
 2   1026    https://leetcode.com/problems/maximum-difference-between-node-and-ancestor/
 2   1305    https://leetcode.com/problems/all-elements-in-two-binary-search-trees/
